refac: Refactor app and create new component PostsPage
